Remove runtime_manager
Removed runtime manager and brought its code back to jit_manager
